Refine job manager
