+++ FhirToDataLake/src/Microsoft.Health.Fhir.Synapse.Core/Jobs/IJobStore.cs
@@ -14,19 +14,20 @@ namespace Microsoft.Health.Fhir.Synapse.Core.Jobs
     public interface IJobStore
     {
         /// <summary>
-        ///  List all active jobs from job store.
+        ///  Acquire an active job from job store.
+        ///  If no active job found, will create a new job trigger and persist to job store.
         /// </summary>
         /// <param name="cancellationToken">cancellation token.</param>
         /// <returns>job lists.</returns>
-        public Task<IEnumerable<Job>> GetActiveJobsAsync(CancellationToken cancellationToken = default);
+        public Task<Job> AcquireJobAsync(CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Update a job in store.
         /// </summary>
         /// <param name="job">the input job.</param>
         /// <param name="cancellationToken">the input cancellationToken.</param>
-        /// <returns>Updated job.</returns>
-        public Task<bool> UpdateJobAsync(Job job, CancellationToken cancellationToken = default);
+        /// <returns>Completed task.</returns>
+        public Task UpdateJobAsync(Job job, CancellationToken cancellationToken = default);
 
         /// <summary>
         /// Complete the running job.
@@ -34,36 +35,7 @@ namespace Microsoft.Health.Fhir.Synapse.Core.Jobs
         /// </summary>
         /// <param name="job">job to complete.</param>
         /// <param name="cancellationToken">cancellation token.</param>
-        /// <returns>Completed job.</returns>
-        public Task<bool> CompleteJobAsync(Job job, CancellationToken cancellationToken = default);
-
-        /// <summary>
-        /// Acquire job lock in job store.
-        /// </summary>
-        /// <param name="cancellationToken">cancellation token.</param>
-        /// <returns>operation result.</returns>
-        public Task<bool> AcquireJobLock(CancellationToken cancellationToken = default);
-
-        /// <summary>
-        /// Release job lock in job store.
-        /// </summary>
-        /// <param name="cancellationToken">cancellation token.</param>
-        /// <returns>operation result.</returns>
-        public Task<bool> ReleaseJobLock(CancellationToken cancellationToken = default);
-
-        /// <summary>
-        /// Get scheduler metadata from job store.
-        /// </summary>
-        /// <param name="cancellationToken">cancellation token.</param>
-        /// <returns>SchedulerSetting object, return null if not exists.</returns>
-        public Task<SchedulerMetadata> GetSchedulerMetadata(CancellationToken cancellationToken = default);
-
-        /// <summary>
-        /// Commit job data from staging folder to result folder.
-        /// </summary>
-        /// <param name="job">input job object.</param>
-        /// <param name="cancellationToken">cancellation token.</param>
-        /// <returns>completed task.</returns>
-        public Task CommitJobDataAsync(Job job, CancellationToken cancellationToken = default);
+        /// <returns>Completed task.</returns>
+        public Task CompleteJobAsync(Job job, CancellationToken cancellationToken = default);
     }
 }
